feat(was): 하입 스코어 표시·정렬을 소수 컬럼으로 전환
/v1/contents·/v1/influencers?sort=hype의 hypeScore 표시·정렬을 정수
hype_score/avg_hype_score에서 신설 소수 컬럼 hype_score_precise/
avg_hype_score_precise로 옮긴다. ContentCard.hypeScore(Long→BigDecimal)·
InfluencerCard.hypeScore(Integer→BigDecimal) 타입 변경 — 자리수 조정은
프론트 몫이라 자르기만 하고 그대로 응답에 싣는다.

ContentCardRow·CardRow에 hypeScorePrecise/avgHypeScorePrecise 필드 추가
(구 hypeScore/avgHypeScore 필드는 유지 — SELECT에서 여전히 읽지만 카드
조립에는 이제 안 쓴다). orderBy()의 hype 정렬 키도 각각 hype_score_precise·
avg_hype_score_precise로 교체.

테스트: DDL 픽스처에 hype_score_precise/avg_hype_score_precise 컬럼 추가,
직접 record 생성 지점 전부 갱신, 정수 동점·소수 비동점 회귀(V1ContentRepositoryTest
z1/a1, V1InfluencerDiscoveryRepositoryTest hype_정렬은_avg_hype_score_정수가_같아도_precise로_순서를_가른다)
추가.

+++ was/src/main/java/com/celfit/was/v1/influencer/InfluencerCard.java
@@ -8,13 +8,15 @@ import java.util.List;
  * id는 handle 그대로(6.4 확정 준용). email은 biography 정규식 파싱(V46, 스펙
  * 2026-07-30-influencer-email-from-bio) — 매치 없으면 null.
  * effectiveFollowers·avgViews·er는 산출 불가(ER·릴스 없음)면 null, bio·tagline 부재는 빈 문자열.
- * hypeScore: 계정 하입 스코어(최근창 콘텐츠 hype_score 단순 평균, 0~100) — 점수 가능 콘텐츠 없으면 null
- * (스펙 2026-07-29-influencer-avg-hype-score).
+ * hypeScore: 계정 하입 스코어(최근창 콘텐츠 출력 매핑 반영 평균, 0~100) — 점수 가능 콘텐츠 없으면
+ * null(스펙 2026-07-29-influencer-avg-hype-score). 2026-07-30부터 소수(BigDecimal) —
+ * account_summaries.avg_hype_score_precise를 그대로 싣는다(스펙
+ * 2026-07-30-hype-score-v3-decay-after-mapping-design.md §10). 자리수 조정은 프론트 몫.
  */
 public record InfluencerCard(String id, String handle, String displayName, String profileImageUrl,
 		Long followers, Long effectiveFollowers, Long postsCount, Long followingCount, String bio,
 		String email, String tagline, BigDecimal reachMultiplier, BigDecimal er, Long avgViews,
-		Long avgLikes, Long avgComments, Integer hypeScore, Long sponsoredCount,
+		Long avgLikes, Long avgComments, BigDecimal hypeScore, Long sponsoredCount,
 		List<String> collaboratedBrands, List<CategoryShare> categoryShares,
 		List<RecentThumb> recentThumbs) {
